+++ src/main/java/com/github/singond/music/Chord.java
@@ -16,6 +16,7 @@
 
 package com.github.singond.music;
 
+import java.util.Iterator;
 import java.util.List;
 
 /**
@@ -29,7 +30,7 @@ import java.util.List;
  *
  * @author Singon
  */
-public interface Chord {
+public interface Chord extends Iterable<PitchClass> {
 
 	/**
 	 * Returns the interval structure of this chord.
@@ -81,58 +82,15 @@ public interface Chord {
 	Interval span();
 
 	/**
-	 * Checks whether this chord has any non-zero inversions,
-	 * ie. whether {{@link #invert} can be called without throwing
-	 * {@code UnsupportedOperationException}.
-	 *
-	 * @return {@code true}, if calling {@link #invert} will not result
-	 *         in {@code UnsupportedOperationException}
-	 */
-	boolean invertible();
-
-	/**
-	 * Returns the <em>n</em>-th inversion of this chord, where <em>n</em>
-	 * is a number between 0 (inclusive) and the chord size (exclusive).
-	 * The <em>n</em>-th inversion of a chord is such a rearrangement of
-	 * its notes that the <em>n</em>-th note (counting from the root,
-	 * which is assigned the index 0) is in the bass.
-	 * <p>
-	 * Apart from this requirement for the bass note, the exact positions
-	 * of remaining notes are un-specified, and the implementations are
-	 * free to choose the exact form of the inversion.
-	 *
-	 * @param n the inversion number <em>n</em> (see above)
-	 * @return an inversion of this chord with the {@code n}-th note
-	 *         in the bass
-	 * @throws IndexOutOfBoundsException if {@code n} is not between
-	 *         0 (inclusive) and {@code size()} (exclusive)
-	 * @throws UnsupportedOperationException if this chord has no inversions
-	 */
-	Chord invert(int n);
-
-	/**
 	 * Returns the number of inversion of this chord.
 	 * If the notes in the chord are indexed from 0 starting with the
 	 * <em>root</em> in ascending order and wrapping down to base upon
 	 * reaching the top note, the inversion number is the index
 	 * of the <em>bass</em> note.
-	 * <p>
-	 * Equivalently, this is the argument to {@link #invert(int)} needed
-	 * to produce this chord type.
 	 *
 	 * @return the index of the bass note
 	 */
 	int inversion();
-
-	/**
-	 * Returns the root position of this chord type.
-	 * The root position of a chord is the position in which the root
-	 * note is equal to the bass note.
-	 * This is equivalent to calling {@code invert(0)}.
-	 *
-	 * @return the root position of this chord
-	 */
-	Chord rootPosition();
 
 	/**
 	 * Returns the chord type of this chord.
@@ -140,4 +98,16 @@ public interface Chord {
 	 * @return the chord type
 	 */
 	ChordType type();
+
+	/**
+	 * Returns an iterator over the pitch classes in this chord in ascending
+	 * order starting from the bass note.
+	 * The returned iterator does not allow modifications to this chord.
+	 * Attempts to modify it will result in an
+	 * {@code UnsupportedOperationException}.
+	 *
+	 * @return {@inheritDoc}
+	 */
+	@Override
+	Iterator<PitchClass> iterator();
 }